Register signal ops for op set 17.
Also:
* de-duplicate get_scalar_value_from_tensor
* fix some bugs that caused compilation errors with the experimental
  ops. Tested with `build.sh --ms_experimental`

[ghstack-poisoned]
