+++ onnxruntime/contrib_ops/cpu/signal/window_functions.h
@@ -3,18 +3,10 @@
 
 #ifdef BUILD_MS_EXPERIMENTAL_OPS
 
+#include "core/providers/cpu/signal/window_functions.h"
+
 namespace onnxruntime {
 namespace contrib {
-
-class VariableOutputDataTypeBase : public OpKernel {
- protected:
-  onnx::TensorProto_DataType data_type_;
-
- public:
-  VariableOutputDataTypeBase(const OpKernelInfo& info) : OpKernel(info) {
-    data_type_ = static_cast<onnx::TensorProto_DataType>(info.GetAttrOrDefault<int64_t>("output_datatype", onnx::TensorProto_DataType::TensorProto_DataType_FLOAT));
-  }
-};
 
 class HannWindow final : public VariableOutputDataTypeBase {
  public:
@@ -33,13 +25,6 @@ class HammingWindow final : public VariableOutputDataTypeBase {
 class BlackmanWindow final : public VariableOutputDataTypeBase {
  public:
   explicit BlackmanWindow(const OpKernelInfo& info) : VariableOutputDataTypeBase(info) {
-  }
-  Status Compute(OpKernelContext* ctx) const override;
-};
-
-class MelWeightMatrix final : public VariableOutputDataTypeBase {
- public:
-  explicit MelWeightMatrix(const OpKernelInfo& info) : VariableOutputDataTypeBase(info) {
   }
   Status Compute(OpKernelContext* ctx) const override;
 };
