+++ onnxruntime/core/providers/cpu/signal/dft.h
@@ -1,15 +1,16 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#ifdef BUILD_MS_EXPERIMENTAL_OPS
+#include "core/common/common.h"
+#include "core/framework/op_kernel.h"
 
 namespace onnxruntime {
-namespace contrib {
 
 class DFT final : public OpKernel {
   bool is_onesided_ = true;
   int64_t axis_ = 0;
   bool is_inverse_ = false;
+
  public:
   explicit DFT(const OpKernelInfo& info) : OpKernel(info) {
     is_onesided_ = static_cast<bool>(info.GetAttrOrDefault<int64_t>("onesided", 0));
@@ -21,6 +22,7 @@ class DFT final : public OpKernel {
 
 class IDFT final : public OpKernel {
   int64_t axis_ = 0;
+
  public:
   explicit IDFT(const OpKernelInfo& info) : OpKernel(info) {
     axis_ = info.GetAttrOrDefault<int64_t>("axis", 0);
@@ -30,6 +32,7 @@ class IDFT final : public OpKernel {
 
 class STFT final : public OpKernel {
   bool is_onesided_ = true;
+
  public:
   explicit STFT(const OpKernelInfo& info) : OpKernel(info) {
     is_onesided_ = static_cast<bool>(info.GetAttrOrDefault<int64_t>("onesided", 1));
@@ -37,7 +40,4 @@ class STFT final : public OpKernel {
   Status Compute(OpKernelContext* ctx) const override;
 };
 
-}  // namespace contrib
 }  // namespace onnxruntime
-
-#endif
